Made Shiftable ABC
