+++ permuta/__init__.py
@@ -1,5 +1,6 @@
 from .pattern import Pattern
 from .rotatable import Rotatable
+from .shiftable import Shiftable
 from .permutation import Permutation
 from .permutations import Permutations
 from .mesh_pattern import MeshPattern
